User can specify the color range for leaves and inner nodes.
There is one shader for each city used for all materials of leaves and
inner nodes in that city. The shader is parameterized by the culling area (portalMin
and portalMax). This way we can cull objects of different cities separately.

+++ Assets/SEE/GameObjects/NodeFactories/BuildingFactory.cs
@@ -18,9 +18,9 @@ namespace SEE.GO
         /// The number of building types offered.
         /// </summary>
         /// <returns>number of building types offered</returns>
-        public override int NumberOfStyles()
+        public override uint NumberOfStyles()
         {
-            return prefabFiles.Length;
+            return (uint)prefabFiles.Length;
         }
 
         /// <summary>
